Rework the entire setup
