--- .idea/modules.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="ProjectModuleManager">
-    <modules>
-      <module fileurl="file://$PROJECT_DIR$/.idea/modules/opus-jni.iml" filepath="$PROJECT_DIR$/.idea/modules/opus-jni.iml" />
-      <module fileurl="file://$PROJECT_DIR$/.idea/modules/opus-jni.main.iml" filepath="$PROJECT_DIR$/.idea/modules/opus-jni.main.iml" />
-      <module fileurl="file://$PROJECT_DIR$/.idea/modules/opus-jni.test.iml" filepath="$PROJECT_DIR$/.idea/modules/opus-jni.test.iml" />
-    </modules>
-  </component>
-</project>
