+++ build.gradle
@@ -1,68 +1,58 @@
-plugins {
-    id 'java'
-    id 'c'
-}
+import org.apache.tools.ant.taskdefs.condition.Os
 
 group 'net.labymod'
 version '1.0'
 
-repositories {
-    mavenCentral()
+ext.generatedHeadersDir = new File(buildDir, "generatedHeaders")
+if(!ext.generatedHeadersDir.exists() && !ext.generatedHeadersDir.mkdirs()) {
+    throw new GradleException("Failed to create directory ${ext.generatedHeadersDir.absolutePath}")
 }
 
-sourceCompatibility = 1.8
-targetCompatibility = 1.8
-
-dependencies {
-    testCompile group: 'junit', name: 'junit', version: '4.12'
-    compile group: 'org.apache.commons', name: 'commons-io', version: '1.3.2'
-    compile 'net.sourceforge.argparse4j:argparse4j:0.8.1'
+ext.nativeBinariesDir = new File(buildDir, "nativeBinaries")
+if(!ext.nativeBinariesDir.exists() && !ext.nativeBinariesDir.mkdirs()) {
+    throw new GradleException("Failed to create directory ${ext.nativeBinariesDir.absolutePath}")
 }
 
-jar {
-    from {
-        configurations.compile.collect { it.isDirectory() ? it : zipTree(it) }
-    }
-}
+def isExternalBuild = false
 
-
-/*
-
-model {
-    platforms {
-        x64 {
-            architecture "x64"
+if(!project.hasProperty("nativeBinaryExternalDir")) {
+    if (Os.isFamily(Os.FAMILY_WINDOWS)) {
+        // WARNING, this assumes CMake uses visual studio
+        // For GCC or Clang the output should be $nativeBinariesDir/libopus-jni-natibe.dll
+        ext.nativeBinaries = Arrays.asList(new File(nativeBinariesDir, "Debug/opus-jni-native.dll"))
+    } else if (Os.isFamily(Os.FAMILY_MAC)) {
+        ext.nativeBinaries = Arrays.asList(new File(nativeBinariesDir, "libopus-jni-native.dylib"))
+    } else if (Os.isFamily(Os.FAMILY_UNIX)) {
+        ext.nativeBinaries = Arrays.asList(new File(nativeBinariesDir, "libopus-jni-native.so"))
+    } else {
+        if (!project.hasProperty("nativeBinaryLibrary")) {
+            throw new GradleException("Unable to determine native output library, " +
+                    "pass -PnativeBinaryLibrary=/path/to/native/binary to gradle")
+        } else {
+            ext.nativeBinaries = [new File(project.property("nativeBinaryLibrary").toString())]
         }
     }
-    repositories {
-        libs(PrebuiltLibraries) {
-            opus {
-                binaries.withType(SharedLibraryBinary) {
-                    sharedLibraryFile = file('/usr/lib/libopus.so')
-                }
-            }
-        }
+} else {
+    def nativeBinaryExternalDir = new File(project.property("nativeBinaryExternalDir").toString())
+    if(!ext.nativeBinaryExternalDir.exists()) {
+        throw new GradleException("nativeBinaryExternalDir ${nativeBinaryExternalDir.absolutePath} does not exist")
     }
 
-    components {
-        opus_jni(NativeLibrarySpec) {
-            sources {
-                c {
-                    lib library: 'opus', linkage: 'shared'
-                }
-            }
-            binaries.all {
-                if (targetPlatform.operatingSystem.linux) {
-                    cCompiler.args '-I', "${org.gradle.internal.jvm.Jvm.current().javaHome}/include"
-                    cCompiler.args '-I', "${org.gradle.internal.jvm.Jvm.current().javaHome}/include/linux"
-                    cCompiler.args '-I', "/usr/include/opus"
-                    cCompiler.args '-D_FILE_OFFSET_BITS=64'
-                }
-            }
+    ext.nativeBinaries = new ArrayList<File>()
+
+    for(def file in nativeBinaryExternalDir.listFiles()) {
+        ext.nativeBinaries.add(file)
+    }
+
+    isExternalBuild = true
+}
+
+project(":opus-jni-java").afterEvaluate { javaProject ->
+    project("opus-jni-native").afterEvaluate { nativeProject ->
+        nativeProject.tasks["build"].dependsOn(javaProject.tasks["generateJniHeaders"])
+
+        if(!isExternalBuild) {
+            javaProject.tasks["processResources"].dependsOn(nativeProject.tasks["build"])
         }
     }
 }
-*/
-
-
-//test.dependsOn 'opus_jniSharedLibrary'
