+++ opus-jni-native/3rdparty/CMakeLists.txt
@@ -1,0 +1,4 @@
+############
+# Add Opus #
+############
+add_subdirectory(opus)
